+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to the "inform-job-support" extension will be documented in 
 ### Added
 - Support Diagnosis to check duplicated IO names for IONAME.DAT and EXIONAME.DAT.
 
+### Changed
+- Delete diagnostic when file is closed.
+
 ## [1.3.0] - 2023-02-27
 ### Added
 - Support Hover to show IO name for JBI files.