+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to the "inform-job-support" extension will be documented in 
 
 ## [Unreleased]
 ### Added
-- Support Diagnosis to check duplicated IO names for IONAME.DAT and EXIONAME.DAT.
+- Support Diagnosis to check duplicated IO names for IONAME.DAT.
 - Support Diagnosis to check duplicated Variable names for VARNAME.DAT.
+- Support Code Action to fix duplicated IO names for IONAME.DAT.
 
 ### Changed
 - Delete diagnostic when file is closed.